+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## (2021-11-02)
+- added azure-ts-serverless-www-html
+
 ## (2021-10-18)
 - added azure-go-sqlserver-servervulnerabilityassessment